pyproject.toml
[project]
name = "bot_template"
version = "0.1.0"
description = "This template provides clean start to create bot using aiogram."
authors = [
    {name = "Author Name", email = "arturboyun@gmail.com"},
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram-dialog>=2.2.0",
    "aiogram>=3.13.1",
    "alembic>=1.13.3",
    "asyncpg>=0.30.0",
    "dishka>=1.4.0",
    "fastapi[standard]>=0.115.5",
    "pydantic-settings>=2.6.0",
    "redis>=5.2.0",
    "ruff>=0.7.1",
    "sqlalchemy[asyncio]>=2.0.36",
    "loguru>=0.7.3",
    "sqlmodel>=0.0.24",
    "psycopg[binary]>=3.2.9",
    "fake-useragent>=2.2.0",
    "httpx[http2]>=0.27.2",
    "playwright>=1.53.0",
    "pytz>=2025.2",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.7.1",
]
